♻️ Post, Comments 분리 및 상태 관리 방법 변경 (#83)
